Upgrade @astrojs/prism to a real package, fix component import not working
