--- packages/astro-prism/index.js
@@ -1,1 +1,0 @@
-export { default as Prism } from './Prism.astro';
