+++ packages/astro-prism/package.json
@@ -3,6 +3,7 @@
   "version": "0.6.1",
   "description": "Supports Prism highlighting in Astro projects",
   "author": "withastro",
+  "type": "module",
   "license": "MIT",
   "bugs": "https://github.com/withastro/astro/issues",
   "repository": {
@@ -11,16 +12,27 @@
     "directory": "packages/astro-prism"
   },
   "homepage": "https://astro.build",
-  "main": "index.js",
-  "scripts": {},
-  "exports": {
-    ".": "./index.js",
-    "./internal": "./internal.mjs"
+  "main": "dist/index.js",
+  "scripts": {
+    "build": "astro-scripts build \"src/**/*.ts\" && tsc",
+    "build:ci": "astro-scripts build \"src/**/*.ts\"",
+    "dev": "astro-scripts dev \"src/**/*.ts\""
   },
-  "types": "./internal.d.ts",
-  "keywords": [],
-  "devDependencies": {
+  "exports": {
+    ".": "./dist/index.js",
+    "./dist/internal": "./dist/internal.js",
+    "./Prism.astro": "./Prism.astro"
+  },
+  "keywords": [
+    "astro",
+    "astro-component"
+  ],
+  "dependencies": {
     "prismjs": "^1.28.0"
+  },
+  "devDependencies": {
+    "astro-scripts": "workspace:*",
+    "@types/prismjs": "1.26.0"
   },
   "engines": {
     "node": "^14.18.0 || >=16.12.0"
