+++ packages/astro-prism/src/index.ts
@@ -1,0 +1,2 @@
+// @ts-expect-error
+export { default as Prism } from '../Prism.astro';
