+++ packages/astro/package.json
@@ -124,7 +124,6 @@
     "postcss": "^8.4.14",
     "postcss-load-config": "^3.1.4",
     "preferred-pm": "^3.0.3",
-    "prismjs": "^1.28.0",
     "prompts": "^2.4.2",
     "recast": "^0.20.5",
     "rehype": "^12.0.1",
