+++ packages/integrations/mdx/package.json
@@ -35,7 +35,6 @@
     "@mdx-js/rollup": "^2.1.1",
     "es-module-lexer": "^0.10.5",
     "gray-matter": "^4.0.3",
-    "prismjs": "^1.28.0",
     "rehype-raw": "^6.1.1",
     "remark-frontmatter": "^4.0.1",
     "remark-gfm": "^3.0.1",
