+++ packages/markdown/remark/package.json
@@ -36,7 +36,6 @@
     "micromark-extension-mdx-expression": "^1.0.3",
     "micromark-extension-mdx-md": "^1.0.0",
     "micromark-util-combine-extensions": "^1.0.0",
-    "prismjs": "^1.28.0",
     "rehype-raw": "^6.1.1",
     "rehype-stringify": "^9.0.3",
     "remark-gfm": "^3.0.1",
@@ -55,7 +54,6 @@
     "@types/hast": "^2.3.4",
     "@types/mdast": "^3.0.10",
     "@types/mocha": "^9.1.1",
-    "@types/prismjs": "^1.26.0",
     "@types/unist": "^2.0.6",
     "astro-scripts": "workspace:*",
     "chai": "^4.3.6",
